refactor: extract docs business logic from commands/ to core/
Move all business logic out of commands/docs.rs into two new core modules:

- core/codebase_map.rs (1073 lines): CodebaseMap builder, markdown renderer,
  source directory detection, class prefix splitting, module name derivation.
  Replaces the hand-rolled collect_fingerprints_recursive with
  codebase_scan::walk_files (the canonical file walker already used by
  refactor rename/transform/move_items and symbol_graph).

- core/docs.rs (680 lines): GenerateSpec/GenerateResult types, spec-based
  file generation with sibling section inference, audit-to-docs rendering
  with extension DocTarget templates, replace_or_append_section, and all
  associated tests.

commands/docs.rs is now 253 lines of thin CLI glue (was 2007): parse args,
dispatch to core, format output. No business logic remains in commands/.
