+++ src/core/mod.rs
@@ -4,11 +4,13 @@ pub mod config;
 pub mod api;
 pub mod auth;
 pub mod cli_tool;
+pub mod codebase_map;
 pub mod code_audit;
 pub mod component;
 pub mod context;
 pub mod db;
 pub mod deploy;
+pub mod docs;
 pub mod engine;
 pub mod error;
 pub mod extension;
